+++ src/es/unileon/ulebank/brokerage/pack/InvestmentFundPack.java
@@ -4,7 +4,7 @@ import es.unileon.ulebank.brokerage.buyable.Buyable;
 
 public class InvestmentFundPack extends Pack {
 
-    public InvestmentFundPack(Buyable product, int amount) {
+    public InvestmentFundPack(Buyable product, long amount) {
         super(product, amount);
     }
 }
